Read MCMC chain associated with a BayesSpace clustering or enhancement

BayesSpace stores the MCMC chain associated with a clustering or enhancement on disk in an HDF5 file. The mcmcChain() function reads any parameters specified by the user into a coda::mcmc object compatible with TidyBayes.

mcmcChain(sce, params = NULL)

removeChain(sce)

Arguments

sce
SingleCellExperiment with a file path stored in its metadata.

params
List of model parameters to read

Value

Returns an mcmc object containing the values of the requested parameters over the constructed chain.

Details

To interact with the HDF5 file directly, obtain the filename from the SingleCellExperiment's metadata: metadata(sce)$chain.h5. Each parameter is stored as a separate dataset in the file, and is represented as a matrix of size (n_iterations x n_parameter_indices).

Examples

set.seed(149)
sce <- exampleSCE()
sce <- spatialCluster(sce, 7, nrep=200, burn.in=20, save.chain=TRUE)

#> Neighbors were identified for 96 out of 96 spots.

#> Fitting model...

#> Calculating labels using iterations 20 through 200.

chain <- mcmcChain(sce)
removeChain(sce)