Spatial clustering

Cluster a spatial expression dataset.

spatialCluster(
  sce,
  q,
  use.dimred = "PCA",
  d = 15,
  platform = c("Visium", "ST"),
  init = NULL,
  init.method = c("mclust", "kmeans"),
  model = c("t", "normal"),
  precision = c("equal", "variable"),
  nrep = 50000,
  gamma = 3,
  mu0 = NULL,
  lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  save.chain = FALSE,
  chain.fname = NULL
)

Arguments

sce
A SingleCellExperiment object containing the spatial data.

q
The number of clusters.

use.dimred
Name of a reduced dimensionality result in reducedDims(sce). If provided, cluster on these features directly.

d
Number of top principal components to use when clustering.

platform
Spatial transcriptomic platform. Specify 'Visium' for hex lattice geometry or 'ST' for square lattice geometry.

init
Initial cluster assignments for spots.

init.method
If init is not provided, cluster the top d PCs with this method to obtain initial cluster assignments.

model
Error model. ('normal' or 't')

precision
Covariance structure. ('equal' or 'variable' for EEE and VVV covariance models, respectively.)

nrep
The number of MCMC iterations.

gamma
Smoothing parameter. (Values in range of 1-3 seem to work well.)

mu0
Prior mean hyperparameter for mu. If not provided, mu0 is set to the mean of PCs over all spots.

lambda0
Prior precision hyperparam for mu. If not provided, lambda0 is set to a diagonal matrix \(0.01 I\).

alpha
Hyperparameter for Wishart distributed precision lambda.

beta
Hyperparameter for Wishart distributed precision lambda.

save.chain
If true, save the MCMC chain to an HDF5 file.

chain.fname
File path for saved chain. Tempfile used if not provided.

Value

Returns a modified sce with cluster assignments stored in colData under the name spatial.cluster.

Details

TODO describe method in detail TODO add details or link to mcmcChain

Examples

set.seed(149)
sce <- exampleSCE()
sce <- spatialCluster(sce, 7, nrep=200)

#> Neighbors were identified for 96 out of 96 spots.

#> Fitting model...

#> Calculating labels using iterations 2 through 200