Enhance spot resolution

Backend calls iterate_deconv(), written in Rcpp. Inputs are the same as spatialCluster() except you have to specify xdist and ydist instead of total dist...(maybe would be better to change spatialCluster to match this)

spatialEnhance(
  sce,
  q,
  platform = c("Visium", "ST"),
  use.dimred = "PCA",
  d = 15,
  init = NULL,
  init.method = c("spatialCluster", "mclust", "kmeans"),
  model = c("t", "normal"),
  nrep = 2e+05,
  gamma = 3,
  mu0 = NULL,
  lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  save.chain = FALSE,
  chain.fname = NULL,
  burn.in = 10000,
  jitter_scale = 5,
  jitter_prior = 0.3,
  verbose = FALSE
)

Arguments

sce
A SingleCellExperiment object containing the spatial data.

q
The number of clusters.

platform
Spatial transcriptomic platform. Specify 'Visium' for hex lattice geometry or 'ST' for square lattice geometry.

use.dimred
Name of a reduced dimensionality result in reducedDims(sce). If provided, cluster on these features directly.

d
Number of top principal components to use when clustering.

init
Initial cluster assignments for spots.

init.method
If init is not provided, cluster the top d PCs with this method to obtain initial cluster assignments.

model
Error model. ('normal' or 't')

nrep
The number of MCMC iterations.

gamma
Smoothing parameter. (Values in range of 1-3 seem to work well.)

mu0
Prior mean hyperparameter for mu. If not provided, mu0 is set to the mean of PCs over all spots.

lambda0
Prior precision hyperparam for mu. If not provided, lambda0 is set to a diagonal matrix \(0.01 I\).

alpha
Hyperparameter for Wishart distributed precision lambda.

beta
Hyperparameter for Wishart distributed precision lambda.

save.chain
If true, save the MCMC chain to an HDF5 file.

chain.fname
File path for saved chain. Tempfile used if not provided.

burn.in
Number of iterations to exclude as burn-in period. The MCMC iterations are currently thinned to every 100; accordingly burn.in is rounded down to the nearest multiple of 100.

jitter_scale
Controls the amount of jittering. Small amounts of jittering are more likely to be accepted but result in exploring the space more slowly. We suggest tuning jitter_scale so that Ychange is on average around 30%.

jitter_prior
Scale factor for the prior variance, parameterized as the proportion (default = 0.3) of the mean variance of the PCs. We suggest making jitter_prior smaller if the jittered values are not expected to vary much from the overall mean of the spot.

verbose
Log progress to stderr.

Value

Returns a new SingleCellExperiment object. By default, the assays of this object are empty, and the enhanced resolution PCs are stored as a reduced dimensionality result accessible with reducedDim(sce, 'PCA').

Details

TODO: add details on subspot coordinates and linking to parent spots

Examples

set.seed(149)
sce <- exampleSCE()
sce <- spatialCluster(sce, 7, nrep=200, burn.in=20)

#> Neighbors were identified for 96 out of 96 spots.

#> Fitting model...

#> Calculating labels using iterations 20 through 200

enhanced <- spatialEnhance(sce, 7, init=sce$spatial.cluster, nrep=200, burn.in=20)

#> Calculating labels using iterations 0 through 200